gym-service-main/requirements.txt
# django
django==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0

# code fomatting
black==22.8.0
isort==5.10.1

# db-tools
dj-database-url==1.0.0
psycopg2-binary==2.8.4

# testing tools
factory-boy==3.2.1
Faker==15.0.0
pytest==7.1.3
pytest-django==4.5.2